+++ zGameServer_NA/GameServer/CallStackTrace.h
@@ -7,18 +7,25 @@
 
 class CCallStackTrace
 {
+
 public:
+
 	CCallStackTrace(char * const szLogID,char * const szLogFileName, char * const szLogDirectoryName);
 	virtual ~CCallStackTrace();
-	int SaveCallStack(BOOL bWriteInFile);
+
+	int  SaveCallStack(BOOL bWriteInFile);
+	//char* __thiscall  GetCallStackLogBuffer();
 
 private:
-	void Output();
-	int TraceStack(BOOL bWriteInFile);
-	DWORD CCallStackTrace::ConvertAddress(HANDLE hProcess,DWORD address, char* output_buffer);
-	static DWORD __stdcall CCallStackTrace::GetModuleBase(HANDLE hProcess, DWORD address);
+
+	void  Output();
+	int  TraceStack(BOOL bWriteInFile);
+	unsigned long  CCallStackTrace::ConvertAddress(HANDLE hProcess,DWORD address, char* output_buffer);
+
+	static unsigned long __stdcall CCallStackTrace::GetModuleBase(HANDLE hProcess, DWORD address);
 
 private:
+
 	FILE * m_fLogFile;	// 4
 	char m_szLogID[260];	// 8
 	char m_szLogFileName[260];	// 10C
@@ -26,6 +33,9 @@ private:
 	char m_szCallFuncName[260];	// 314
 	CRITICAL_SECTION m_critLogToFile;	// 418
 	std::string m_strCallStackLog;	// 430
+
+
 };
 	
+
 #endif
