+++ zGameServer_NA/GameServer/SkillDelay.cpp
@@ -1,6 +1,12 @@
-//GameServer 1.00.90 finished
+// ------------------------------
+// Decompiled by Deathway
+// Date : 2007-03-09
+// ------------------------------
 #include "stdafx.h"
 #include "SkillDelay.h"
+
+// GS-N 0.99.60T Status Completed : in oder version the m_SkillDelay coudl change
+//	GS-N	1.00.18	JPN	0x00485540	-	Completed
 
 CSkillDelay::CSkillDelay()
 {
@@ -21,7 +27,7 @@ void CSkillDelay::Init()
 
 
 #pragma warning ( disable : 4101 )
-int CSkillDelay::Check(BYTE skill)
+int CSkillDelay::Check(WORD skill)
 {
 	int skilldelaytime;
 	DWORD dwtime;
